Clarify some glossary terms
Update glossary terms related to "command" and submission to the queue.
In my mind, a "command" and a "command group" aren't submitted to a
queue.  Rather, these are nodes in the dependency graph.  A SYCL
application submits a "command group function object" to a queue, and
this causes the implementation to add a "command group" to the
dependency graph.  By clarifying the meaning of these terms, I think we
can start using them more consistently in the rest of the spec.