Make swizzle view members const
We discovered after some implemenation experience that the definitions
proposed in the early commits of this PR don't support the desired use
cases.  A simple example is:

```
sycl::vec<int, 4> v;
v.swizzle<1>()++;
```

This won't compile because the declaration of `operator++` takes a
mutable reference to the swizzle object:

```
friend vec<DataT, NumElements> operator++(__swizzled_vec__& sv, int);
```

C++ language rules prohibit binding the temporary object returned by
`v.swizzle<>()` to the mutable l-value reference parameter to
`operator++`.

To fix this, we changed all the swizzle operators to take a `const`
reference to the swizzle object.  This makes sense because the swizzle
class is a "view" over the underlying `vec`, and member functions on a
view class are normally `const`, even for functions that mutate the
object that underlies the view.  The logic here is that the view itself
is not modified, only the object referenced by the view.

This raises a new problem, though, because we don't want to allow code
to change a `const vec` via a swizzle.  For example, code like this
should result in a diagnostic:

```
const sycl::vec<int, 4> v;
v.swizzle<1>()++;
```

To fix this, we replace `__swizzled_vec__` with two classes:
`__writeable_swizzle__` and `__const_swizzle__`.  We use
`__const_swizzle__` for swizzles of `const vec`, and this class
provides only the non-mutating operators.